requirements.txt
openai==0.27.0
numpy==1.24.3
sounddevice==0.4.6
python-dotenv==0.19.2
PyQt6==6.5.0
PyQt6-Qt6==6.5.0
PyQt6-sip==13.5.1